+++ source/components/splitter/splitter.js
@@ -51,6 +51,8 @@ var Splitter = {
     _create: function(){
         var element = this.element, o = this.options;
 
+        Metro.checkRuntime(element, "splitter");
+
         this._createStructure();
         this._createEvents();